working on mp3 downloads

+++ chorus.py
@@ -46,7 +46,7 @@ def empty_folder(*args):
         for f in os.listdir(folder):
             file_path = os.path.join(folder, f)
             try:
-                if os.path.isfile(file_path):
+                if os.path.isfile(file_path) and file_path.endswith('.pdf'):
                     os.unlink(file_path)
             except Exception, e:
                 print e
@@ -55,6 +55,7 @@ def parse_page(concert_id, session):
     '''parses page and returns list of music links (.pdf) and list of video links (.mp4)'''
     sheet_music_links = []
     video_links = []
+    recording_links = []
     url = 'http://nycgmc.groupanizer.com/music'
     cookies = login(session, username, password)
     form_build_id, form_token = get_form_params(session, url, cookies)
@@ -82,11 +83,11 @@ def parse_page(concert_id, session):
         for music_link in song_page_soup.find_all('a'):
             if music_link.get('href').endswith('.pdf'):
                 sheet_music_links.append((music_link.text, music_link.get('href')))
-            else:
-                if music_link.get('href').find('youtu') != -1: #finds youtube and youtu.be
-                    video_links.append(music_link.get('href'))
-
-    return sheet_music_links, video_links
+            elif music_link.get('href').find('youtu') != -1: #finds youtube and youtu.be
+                video_links.append(music_link.get('href'))
+            elif music_link.get('href').endswith('.mp3'):
+                recording_links.append((music_link.text, music_link.get('href')))   
+    return sheet_music_links, video_links, recording_links
 
 def write_sheet_music_to_file(session, sheet_music_links, directory_str):
     '''takes as input list of music links and writes all sheet music to specified directory_str
@@ -114,6 +115,37 @@ def write_videos_to_file(session, video_links, directory_str):
         if not isfile(fullpath):
             video.download(directory_str)
 
+def process_recording_links(session, recording_links, directory_str):
+    #is not checking isfile in correct place
+    tenorI_dir = directory_str + 'tenorI/'
+    tenorII_dir = directory_str + 'tenorII/'
+    baritone_dir = directory_str + 'baritone/'
+    bass_dir = directory_str + 'bass/'
+    for recording_link in recording_links:
+        fullpath = '{0}{1}'.format(directory_str, recording_link[0].replace('/', '_'))
+        if not isfile(fullpath):
+            if fullpath.lower().find('full') != -1:
+                write_recording_to_file(session, recording_link, tenorI_dir)
+                write_recording_to_file(session, recording_link, tenorII_dir)
+                write_recording_to_file(session, recording_link, baritone_dir)
+                write_recording_to_file(session, recording_link, bass_dir)
+            elif fullpath.lower().find('tenor 1')!=-1:
+                write_recording_to_file(session, recording_link, tenorI_dir)
+            elif fullpath.lower().find('tenor 2')!=-1:
+               write_recording_to_file(session, recording_link, tenorII_dir)
+            elif fullpath.lower().find('baritone')!=-1:
+                write_recording_to_file(session, recording_link, baritone_dir)
+            elif fullpath.lower().find('bass')!=-1:
+                write_recording_to_file(session, recording_link, bass_dir)        
+
+def write_recording_to_file(session, link, directory_str):
+    new_filename = link[0].replace('/', '_')
+    f = open(directory_str + new_filename, 'wb')
+    f.write(session.get(link[1]))
+    f.close()
+    return
+
+
 def unshorten_url(url):
     parsed = urlparse.urlparse(url)
     h = httplib.HTTPConnection(parsed.netloc)
@@ -133,10 +165,12 @@ def main():
         directory_str = '/Users/paulnichols/Dropbox/chorus_music/'
     sheet_music_directory = directory_str + 'sheet_music/'
     choralography_directory = directory_str + 'choralography/'
+    recording_directory = directory_str + 'recordings/'
     empty_folder(directory_str, sheet_music_directory, choralography_directory)
-    sheet_music_links, video_links = parse_page(concert_id, session)
-    write_sheet_music_to_file(session, sheet_music_links, sheet_music_directory)
-    write_videos_to_file(session, video_links, choralography_directory)
+    sheet_music_links, video_links, recording_links = parse_page(concert_id, session)
+    # write_sheet_music_to_file(session, sheet_music_links, sheet_music_directory)
+    # write_videos_to_file(session, video_links, choralography_directory)
+    process_recording_links(session, recording_links, recording_directory)
     print 'Finished!'
 
 main()
